refactor: 행정구역 삭제 기능 관련 메서드 및 클래스 삭제(#44)
* presentation 계층 삭제 (RegionController, DeleteRegionRequest)
* RegionFacade 삭제
* RegionRepository, RegionJpaRepository, RegionRepositoryImpl: deleteByAddress 메서드 삭제
* RegionStore, RegionRedisRepository: deleteByAddress 메서드 삭제
